ai_travel_trip_planner/requirements.txt
# Core AI frameworks
agno>=0.9.0
crewai>=0.28.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0

# Google AI
google-generativeai>=0.5.0

# Search tools
duckduckgo-search>=5.0.0
serpapi>=0.1.5

# Streamlit UI
streamlit>=1.35.0
streamlit-extras>=0.4.0

# HTTP & Async
httpx>=0.27.0
aiohttp>=3.9.0
requests>=2.31.0

# Data & Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.26.0

# PDF generation
fpdf2>=2.7.0
markdown>=3.6
